requirements.txt
exif==1.6.*
pillow==10.1.*
python-dateutil==2.8.*
